requirements.txt
numpy
opencv-python
streamlit
Pillow
matplotlib
tensorflow
Keras